## Summary
change decimals to be represented as u8 instead of u64 in native types.
this is because geth uses u8 for decimals, so u64 could have resulted in
truncation issues. decimals shouldn't be larger than u8 anyways.

## Background
fixes zellic audit issue 3.1 "Truncation and precision-loss issue
between Astria and Geth during data transfer".

## Changes
- change decimals to be represented as u8 instead of u64 in native types

## Testing
code logic is unchanged.

+++ crates/astria-sequencer/src/connect/market_map/storage/values/market_map.rs
@@ -51,7 +51,7 @@ impl<'a> From<CurrencyPair<'a>> for DomainCurrencyPair {
 #[derive(Debug, BorshSerialize, BorshDeserialize)]
 struct Ticker<'a> {
     currency_pair: CurrencyPair<'a>,
-    decimals: u64,
+    decimals: u8,
     min_provider_count: u64,
     enabled: bool,
     metadata_json: Cow<'a, str>,
